moved various commands below 'fido2 util' (see #6)
